Bug fix, validation, cleaning code.
Server checks if the fields are filled in, it will tell the user to fill them if they are submitted empty.
Changed "Add user to database" to "Register user" on the register page.
Better names for variables.
Cleaned the code.
Added comments for each segment that could be helpful for PHP-beginners.
Renamed files.
Fixed a bug where $fetched_hash was not set.